pyscript/requirements.txt
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==2.0.3
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
openpyxl==3.0.9
pymongo==4.0.2
python-dotenv==0.19.2
Werkzeug==2.0.3
